+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
-## v46.x / v47 (In progress)
+## v47 (2020-09-12)
 - Challenges: New challenge mode "No gifts/store items, all enemy item drops from random pool".
 - Gift box contents: Item ranking slightly adjusted.
+- Music: Allow picking custom tracks to be used in Ancient Cave.
 - Music: Many custom tracks added.
 - PC sprites: Added Ness Pajamas, Prototype Ness, and Prototype Cheese Man.
 ### Bugfixes